split up array into multiple sub arrays
