--- ui/src/pages/namespace/Explorer/Page/poc/schema/primitives/extendedKeyValue.ts
@@ -1,55 +1,0 @@
-import z from "zod";
-
-/**
- * An extended key-value pair that supports multiple data types for the value,
- * including strings, variables, booleans, arrays, and objects.
- */
-const StringSchema = z.object({
-  type: z.literal("string"),
-  value: z.string(),
-});
-
-const VariableSchema = z.object({
-  type: z.literal("variable"),
-  value: z.string(),
-});
-
-const BooleanSchema = z.object({
-  type: z.literal("boolean"),
-  value: z.boolean(),
-});
-
-// for simplifity we don't support nested arrays and objects yet
-const AllowedArrayAndObjectValues = z.union([
-  z.string(),
-  z.number(),
-  z.boolean(),
-]);
-
-const ArraySchema = z.object({
-  type: z.literal("array"),
-  value: z.array(AllowedArrayAndObjectValues),
-});
-
-const ObjectSchema = z.object({
-  type: z.literal("object"),
-  value: z.array(
-    z.object({
-      key: z.string().min(1),
-      value: AllowedArrayAndObjectValues,
-    })
-  ),
-});
-
-const DataType = z.discriminatedUnion("type", [
-  StringSchema,
-  VariableSchema,
-  BooleanSchema,
-  ArraySchema,
-  ObjectSchema,
-]);
-
-export const ExtendedKeyValueSchema = z.object({
-  key: z.string().min(1),
-  value: DataType,
-});
